+++ NEWS.md
@@ -9,6 +9,8 @@ Changes:
 
 3. Adding translation rule for `CAST(... AS DATE)` for SQLite.
 
+4. Added `snakeCaseToCamelCaseNames()` and `camelCaseToSnakeCaseNames()` functions.
+
 
 SqlRender 1.8.3
 ===============